lab-automation-testing/automation-testing/database-tests/requirements.txt
# Database Testing Requirements
# Install with: pip install -r requirements.txt

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1

# MongoDB dependencies
pymongo==4.6.0
motor==3.3.2

# MySQL dependencies
mysql-connector-python==8.2.0

# Alternative MySQL driver (optional)
# PyMySQL==1.1.0

# SQLAlchemy (optional, for ORM testing)
# SQLAlchemy==2.0.23
